requirements.txt
nicegui
opencv-python-headless
plantcv
pandas
numpy
matplotlib
seaborn
Pillow


